Bug Fix
Added an asterisk to the headers so Org file recognized them as such and didn't collapse as part of todo header.  For instance, if I had `* TODO : A Task` and it was above the Research Section, it would stash the research section in the TODO instead of keeping it as it's own section.  Might switch this to a different header at some point but for now it's good to go.